#16 les pages ont ete modifies, les elements default n'apparaissent plus dans la liste

+++ classes/class_Salle.php
@@ -359,45 +359,47 @@
 			
 			$cpt = 0;
 			foreach($liste_id_salles as $idSalle){
-				$Salle = new Salle($idSalle);
-				$couleurFond = ($cpt == 0) ? "fondBlanc" : "fondGris"; $cpt++; $cpt %= 2;
-				$lienInfosSalle = "./index.php?page=infosSalle&amp;idSalle={$Salle->getId()}";
-				if(isset($_GET['idPromotion'])){
-					$lienInfosSalle .= "&amp;idPromotion={$_GET['idPromotion']}";
-				}
-				
-				echo "$tab\t<tr class=\"$couleurFond\">\n";
-				echo "$tab\t\t<td>";
-				echo "<a href=\"$lienInfosSalle\">{$Salle->getNomBatiment()} - {$Salle->getNom()}</a>";
-				echo "</td>\n";
-				echo "$tab\t\t<td>{$Salle->getNomBatiment()}</td>\n";
-				echo "$tab\t\t<td>{$Salle->getNom()}</td>\n";
-				echo "$tab\t\t<td>{$Salle->getCapacite()}</td>\n";
-								
-				foreach($liste_type_salle as $idType_Salle) {					
-					$Type_Salle = new Type_Salle($idType_Salle);
-					$nomType_Salle = $Type_Salle->getNom();
-					if(Type_Salle::appartient_salle_typeSalle($idSalle, $idType_Salle)) 
-						$checked = "checked = \"checked\"" ;
-					else
-						$checked = "";
-					$nomCheckbox = "{$idSalle}_{$nomType_Salle}";
-					echo "$tab\t\t<td><input type=\"checkbox\" name= \"{$idSalle}_{$nomType_Salle}\" value=\"{$idType_Salle}\" onclick=\"appartenance_salle_typeSalle({$idSalle},{$idType_Salle},this)\" style=\"cursor:pointer\" {$checked}></td>\n";
-				}
-			
-				if($administration){
-					$pageModification = "./index.php?page=ajoutSalle&amp;modifier_salle=$idSalle";
-					$pageSuppression = "./index.php?page=ajoutSalle&amp;supprimer_salle=$idSalle";
+				if ($idSalle != 0) {
+					$Salle = new Salle($idSalle);
+					$couleurFond = ($cpt == 0) ? "fondBlanc" : "fondGris"; $cpt++; $cpt %= 2;
+					$lienInfosSalle = "./index.php?page=infosSalle&amp;idSalle={$Salle->getId()}";
 					if(isset($_GET['idPromotion'])){
-						$pageModification .= "&amp;idPromotion={$_GET['idPromotion']}";
-						$pageSuppression .= "&amp;idPromotion={$_GET['idPromotion']}";
+						$lienInfosSalle .= "&amp;idPromotion={$_GET['idPromotion']}";
 					}
+					
+					echo "$tab\t<tr class=\"$couleurFond\">\n";
 					echo "$tab\t\t<td>";
-					echo "<a href=\"$pageModification\"><img alt=\"icone modification\" src=\"../images/modify.png\"></a>";
-					echo "<a href=\"$pageSuppression\" onclick=\"return confirm('Supprimer la salle ?')\"><img alt=\"icone suppression\" src=\"../images/delete.png\" /></a>";
-					echo "</td>";
+					echo "<a href=\"$lienInfosSalle\">{$Salle->getNomBatiment()} - {$Salle->getNom()}</a>";
+					echo "</td>\n";
+					echo "$tab\t\t<td>{$Salle->getNomBatiment()}</td>\n";
+					echo "$tab\t\t<td>{$Salle->getNom()}</td>\n";
+					echo "$tab\t\t<td>{$Salle->getCapacite()}</td>\n";
+									
+					foreach($liste_type_salle as $idType_Salle) {					
+						$Type_Salle = new Type_Salle($idType_Salle);
+						$nomType_Salle = $Type_Salle->getNom();
+						if(Type_Salle::appartient_salle_typeSalle($idSalle, $idType_Salle)) 
+							$checked = "checked = \"checked\"" ;
+						else
+							$checked = "";
+						$nomCheckbox = "{$idSalle}_{$nomType_Salle}";
+						echo "$tab\t\t<td><input type=\"checkbox\" name= \"{$idSalle}_{$nomType_Salle}\" value=\"{$idType_Salle}\" onclick=\"appartenance_salle_typeSalle({$idSalle},{$idType_Salle},this)\" style=\"cursor:pointer\" {$checked}></td>\n";
+					}
+				
+					if($administration){
+						$pageModification = "./index.php?page=ajoutSalle&amp;modifier_salle=$idSalle";
+						$pageSuppression = "./index.php?page=ajoutSalle&amp;supprimer_salle=$idSalle";
+						if(isset($_GET['idPromotion'])){
+							$pageModification .= "&amp;idPromotion={$_GET['idPromotion']}";
+							$pageSuppression .= "&amp;idPromotion={$_GET['idPromotion']}";
+						}
+						echo "$tab\t\t<td>";
+						echo "<a href=\"$pageModification\"><img alt=\"icone modification\" src=\"../images/modify.png\"></a>";
+						echo "<a href=\"$pageSuppression\" onclick=\"return confirm('Supprimer la salle ?')\"><img alt=\"icone suppression\" src=\"../images/delete.png\" /></a>";
+						echo "</td>";
+					}
+					echo "$tab\t</tr>\n";
 				}
-				echo "$tab\t</tr>\n";
 			}
 			
 			echo "$tab</table>\n";
